Did zerosum agent

+++ src/main/java/heuristics/ZeroSumHeuristic.java
@@ -12,7 +12,7 @@ import java.util.Map;
 import java.util.Set;
 
 public class ZeroSumHeuristic {
-    public double evaluate(ZeroSumState state) {
+    public int evaluate(ZeroSumState state) {
         return state.getiSaved()-state.getEnemySaved();
     }
 }
